# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * Architectural Portfolio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  projectCount?: number | null;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
}

export interface ProjectImage {
  id: number;
  projectId: number;
  imageUrl: string;
  isHero: boolean;
  sortOrder: number;
}

export interface ProjectImageInput {
  /** @minLength 1 */
  imageUrl: string;
  isHero?: boolean;
  sortOrder?: number;
}

export interface ProjectImageUpdate {
  imageUrl?: string;
  isHero?: boolean;
  sortOrder?: number;
}

export interface ProjectSummary {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  client?: string | null;
  /** @nullable */
  sector?: string | null;
  status: string;
  published: boolean;
  featured?: boolean;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  serviceId?: number | null;
  /** @nullable */
  heroImage?: string | null;
  /** @nullable */
  year?: string | null;
}

export interface Project {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  client?: string | null;
  /** @nullable */
  sector?: string | null;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  scope?: string | null;
  status: string;
  published: boolean;
  featured?: boolean;
  /** @nullable */
  longDescription?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  /** @nullable */
  serviceId?: number | null;
  /** @nullable */
  heroImage?: string | null;
  /** @nullable */
  year?: string | null;
  images?: ProjectImage[];
}

export interface ProjectInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  slug: string;
  location?: string;
  client?: string;
  sector?: string;
  size?: string;
  scope?: string;
  status: string;
  longDescription?: string;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  serviceId?: number | null;
  year?: string;
  published?: boolean;
  featured?: boolean;
}

export interface PublishToggle {
  published: boolean;
}

export interface MachineryItem {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  longDescription?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  galleryImages?: string | null;
  /** @nullable */
  year?: string | null;
  /** @nullable */
  condition?: string | null;
  published: boolean;
  featured?: boolean;
}

export interface MachineryInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  category?: string;
  description?: string;
  longDescription?: string;
  imageUrl?: string;
  galleryImages?: string;
  year?: string;
  condition?: string;
  published?: boolean;
  featured?: boolean;
}

export interface ServicesItem {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  longDescription?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  galleryImages?: string | null;
  published: boolean;
  featured?: boolean;
  projectCount?: number;
  projects?: ProjectSummary[];
}

export interface ServicesInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  description?: string;
  longDescription?: string;
  imageUrl?: string;
  galleryImages?: string;
  published?: boolean;
  featured?: boolean;
}

export interface AdminPasswordChange {
  /** @minLength 1 */
  currentPassword: string;
  /** @minLength 1 */
  newPassword: string;
}

export interface AdminCredentials {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface AdminSession {
  authenticated: boolean;
}

export interface SiteSettings {
  phone?: string;
  email?: string;
  address?: string;
  city?: string;
  hours?: string;
  heroSubtitle?: string;
  ceoName?: string;
  ceoTitle?: string;
  ceoQuote?: string;
  ceoImage?: string;
}

export type ListProjectsParams = {
published?: boolean;
category_id?: number;
};

export type ListMachineryParams = {
published?: boolean;
};

export type ListServicesParams = {
published?: boolean;
};

export type ChangeAdminPassword200 = {
  success?: boolean;
};