# lib/api-zod/src/generated/types/projectImage.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * Architectural Portfolio API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectImage {
  id: number;
  projectId: number;
  imageUrl: string;
  isHero: boolean;
  sortOrder: number;
}